make workflow in module
