+++ busy_beaver/apps/events/sync_database.py
@@ -1,11 +1,32 @@
+"""Update Events Database Workflow
+
+This workflow pulls a set of events from Meetup and performs the following action
+against the set of future events in the database:
+    - if event is new => create
+    - if event is in the database, but not in the fetched events => delete
+    - if event is in the database => update
+"""
+
 import logging
+import time
 from typing import List, NamedTuple
 
+from busy_beaver.clients import meetup
 from busy_beaver.common.wrappers.meetup import EventDetails
 from busy_beaver.extensions import db
 from busy_beaver.models import Event
 
 logger = logging.getLogger(__name__)
+
+
+def sync_database_with_fetched_events(group_name):
+    fetched_events = meetup.get_events(group_name, count=20)
+
+    current_epoch_time = int(time.time())
+    database_events = Event.query.filter(Event.start_epoch > current_epoch_time).all()
+
+    sync_database = SyncEventDatabase(fetched_events, database_events)
+    sync_database.perform()
 
 
 class PendingTransactions(NamedTuple):
